requirements.txt
streamlit
googlesearch-python
langchain
langchain-openai
python-dotenv